refactor: finish send extraction cleanup
